Player walking sounds and AI tweaks
The player now has footstep noises which vary depending on the surface they're walking on, the AI will react to hearing these noises and run away. In addition the animals now will adjust to slopes on the ground and tilt depending on the floor angle. Also began plugging the AI into an elephant model.