Save the Data
Simple data saving mechanic is now added. Press 1 to save the game, 2 to load it. A "SafariSaveGame" object has been added. This saves the actual data. The save and load macros are in the first person character BP. There is a macro for both saving and loading. You can add items to be saved by making variables in the SafariSaveGame object and then append whatever nodes you need to save to the end of the sequence in the macros.