Photo Gallery
There is now a photo gallery, press G to open and close. All the taken photos are on the left in a scroll box and then photo will appear on the right when clicked on. Note: if no photos are taken, nothing will appear in the gallery. The master volume slider now also works for the in game music.