SS Quiz Use Case
pushing blueprint for safari snapshot quiz. the creator prepares the quiz information by putting in questions, answers, and adds questions if they want. the player answers the questions and if they're correct, the point gets added to the scoreboard and the opposite. if by the end of the game the player scores most of the questions right, the player can exit the quiz. if not, the player can start again.