Animal Indicator
Camera now has an int for export mode. 0 does not export, 1 exports the render target as an exr, 2 exports high res screenshots. It is defaulted to zero to save space and prevent freezing in game. An indicator of how many animals are in the shot will now appear on the bottom left. If there are 2 animals in the picture then 2 squares will appear.